requirements.txt
python-dotenv
openpyxl
psycopg2
progress
fastapi
uvicorn
python-multipart
flask
typing-inspect
